+++ Frameworks/MoEngageCore.framework/Headers/MOCoreConstants.h
@@ -10,25 +10,35 @@
 
 // User Default Keys
 extern NSString* const kUserDefaultSDKSuitName;
-extern NSString* const kUserDefaultKeySDKVersion;
+
+extern NSString* const kUserDefaultsKeySDKVersion;
 extern NSString* const kUserDefaultsKeyAppVersion;
 extern NSString* const kUserDefaultsKeySDKUniqueId;
-extern NSString* const kUserDefaultsKeyShouldRedirectData;
+extern NSString* const kUserDefaultsKeyPushToken;
+extern NSString* const kUserDefaultsKeyMoEngageAppID;
 extern NSString* const kUserDefaultsKeyHasDeviceAddFailed;
 extern NSString* const kUserDefaultsKeyIsSDKEnabled;
+extern NSString* const kUserDefaultsKeyInstallRegistered;
+extern NSString* const kUserDefaultsKeyConfigLastUpdated;
+
+extern NSString* const kUserDefaultsKeyOlderMoEngageAppID;
+extern NSString* const kUserDefaultsKeyOlderSDKUniqueId;
+extern NSString* const kUserDefaultsKeyOlderAppGroupID;
+extern NSString* const kUserDefaultsKeyOlderPushToken;
 
 // Notification Name Keys
 extern NSString* const kNotificationNameUserReset;
 extern NSString* const kNotificationNameShowInApp;
 extern NSString* const kNotificationNameInitNewUser;
 extern NSString* const kNotificationNameUpdateModules;
-extern NSString* const kNotificationNameUserResetOnAppIDChange;
 
 // Date Formats
 extern NSString* const kMODateFormatISO8601;
 extern NSString* const kMODateFormatISO8601WithoutZ;
 extern NSString* const kMODateFormatISO8601MilliSec;
 extern NSString* const kMODateFormatShort;
+extern NSString* const kMOTimeFormatMilliSec;
+extern NSString* const kMOTimeFormatOnlyTimeWithoutZ;
 
 // Integration Constants
 extern NSString* const MoEngage_Cordova_SDK_Version;
@@ -38,7 +48,6 @@ extern NSString* const MoEngage_Xamarin_SDK_Version;
 extern NSString* const MoEngage_Flutter_SDK_Version;
 extern NSString* const MoEngage_Unity_SDK_Version;
 
-// Misc
-extern NSString* const kMoEngageSDKBundleId;
-
+#pragma mark- Other Keys
+extern NSString* const MO_KEYCHAIN_SERVICE;