+++ Frameworks/MoEngageCore.framework/Headers/MOLogger.h
@@ -1,17 +1,24 @@
 //
 //  MOLogger.h
-//  loggerApp
+//  MoEngageCore
 //
-//  Created by Chengappa on 23/10/15.
-//  Copyright © 2015 MoEngage. All rights reserved.
+//  Created by Chengappa on 23/06/20.
+//  Copyright © 2020 MoEngage. All rights reserved.
 //
 
 #import <Foundation/Foundation.h>
+#import "MOLogger.h"
 
+NS_ASSUME_NONNULL_BEGIN
 
-/**
- Enumerator to set Log Levels of SDK for Debugging
- */
+typedef enum {
+    MOLoggerTypeVerbose = 0,
+    MOLoggerTypeDebug ,
+    MOLoggerTypeInfo,
+    MOLoggerTypeWarning,
+    MOLoggerTypeError
+} MOLoggerType;
+
 typedef enum _LogLevel{
     LOG_NONE,
     LOG_ALL,
@@ -19,75 +26,69 @@ typedef enum _LogLevel{
 }LogLevel;
 
 
-/**
- This Enumerator is to set how detailed the logs should be
- */
-typedef enum {
-    MOLoggerVerbosityNone = 0,
-    MOLoggerVerbosityDefault,
-    MOLoggerVerbosityDetailed
-} MOLoggerVerbosity;
+#define MOLogError(st, f, ... )      [[MOLogger sharedInstance] logErrorWithFormatArgs:@[\
+        [[NSString stringWithUTF8String:__FILE__] lastPathComponent],\
+        [NSNumber numberWithInt:__LINE__],\
+        NSStringFromSelector(_cmd),\
+        [NSString stringWithFormat:(f), ##__VA_ARGS__]\
+        ]\
+        andStackTrace:st\
+]
 
 
-/**
- This Enumerator is to set type of log. Currently using MOLoggerSeverityException. Going forward we will use others too.
- */
-typedef enum {
-    MOLoggerSeverityUnset = 0,
-    MOLoggerSeverityDebug,
-    MOLoggerSeverityInfo,
-    MOLoggerSeverityWarn,
-    MOLoggerSeverityException,
-    MOLoggerSeverityError
-} MOLoggerSeverity;
+#define MOLogInfo(f, ... )      [[MOLogger sharedInstance] logWithType: MOLoggerTypeInfo\
+        andFormatArgs:@[\
+        [[NSString stringWithUTF8String:__FILE__] lastPathComponent],\
+        [NSNumber numberWithInt:__LINE__],\
+        NSStringFromSelector(_cmd),\
+        [NSString stringWithFormat:(f), ##__VA_ARGS__]\
+        ]\
+]
 
 
-#define MOSeverityLogDetailed( s, f, ... )      [MOLogger logWithVerbosity:MOLoggerVerbosityDetailed\
-                                                                    severity:s\
-                                                                    formatArgs:@[\
-                                                                        [[NSString stringWithUTF8String:__FILE__] lastPathComponent],\
-                                                                        [NSNumber numberWithInt:__LINE__],\
-                                                                        NSStringFromSelector(_cmd),\
-                                                                        [NSString stringWithFormat:(f), ##__VA_ARGS__]\
-                                                                    ]\
-                                                ]
+#define MOLogDebug(f, ... )       [[MOLogger sharedInstance] logWithType:MOLoggerTypeDebug\
+                    andFormatArgs:@[\
+                        [NSString stringWithFormat:(f), ##__VA_ARGS__]\
+                    ]\
+]
 
-#define MOSeverityLogDetailedWithStackTrace( s,st, f, ... )      [MOLogger logWithVerbosity:MOLoggerVerbosityDetailed\
-                                                                        severity:s\
-                                                                        formatArgs:@[\
-                                                                        [[NSString stringWithUTF8String:__FILE__] lastPathComponent],\
-                                                                        [NSNumber numberWithInt:__LINE__],\
-                                                                        NSStringFromSelector(_cmd),\
-                                                                        [NSString stringWithFormat:(f), ##__VA_ARGS__]\
-                                                                        ]\
-                                                                        andStackTrace:st\
-                                                                ]
+#define MOLogVerbose(f, ... )       [[MOLogger sharedInstance] logWithType:MOLoggerTypeVerbose\
+                    andFormatArgs:@[\
+                        [NSString stringWithFormat:(f), ##__VA_ARGS__]\
+                    ]\
+]
+
+#define MOLogWarning(f, ... )      [[MOLogger sharedInstance] logWithType: MOLoggerTypeWarning\
+        andFormatArgs:@[\
+        [[NSString stringWithUTF8String:__FILE__] lastPathComponent],\
+        [NSNumber numberWithInt:__LINE__],\
+        NSStringFromSelector(_cmd),\
+        [NSString stringWithFormat:(f), ##__VA_ARGS__]\
+        ]\
+]
+
+#define MOLog(f, ... )       [[MOLogger sharedInstance] logWithType:MOLoggerTypeDebug\
+                    andFormatArgs:@[\
+                        [NSString stringWithFormat:(f), ##__VA_ARGS__]\
+                    ]\
+]
 
 
-#define MOSeverityLogDefault( s, f, ... )       [MOLogger logWithVerbosity:MOLoggerVerbosityDefault\
-                                                                    severity:s\
-                                                                    formatArgs:@[\
-                                                                        [NSString stringWithFormat:(f), ##__VA_ARGS__]\
-                                                                    ]\
-                                                ]
-
-#define MOLogDetailed( f, ... )            MOSeverityLogDetailed( MOLoggerSeverityUnset, f, ##__VA_ARGS__ )
-#define MOLogDefault( f, ... )             MOSeverityLogDefault( MOLoggerSeverityUnset, f, ##__VA_ARGS__ )
-
-#define MOLog( f, ... )                    MOLogDefault( f, ##__VA_ARGS__ )
-#define MOSeverityLog( s, f, ... )  MOSeverityLogDefault( s, f, ##__VA_ARGS__ )
-
-#ifdef MOLOGGER_SWIZZLE_NSLOG
-#define NSLog( s, ... )		MOLog( s, ##__VA_ARGS__ )
-#endif
 
 @interface MOLogger : NSObject
 
-+ (void)log:(NSString *)format, ...;
-+ (void)logWithVerbosity:(MOLoggerVerbosity)verbosity severity:(MOLoggerSeverity)severity formatArgs:(NSArray *)args;
-+ (void)logWithVerbosity:(MOLoggerVerbosity)verbosity severity:(MOLoggerSeverity)severity formatArgs:(NSArray *)args andStackTrace:(NSArray<NSString*>*)stackTrace;
+@property (class, assign, readonly) NSInteger maxLogDuration;
+@property (class, assign, readonly) NSInteger minLogDuration;
 
-+ (MOLoggerSeverity)minimumSeverity;
-+(void)debug:(LogLevel)logLevel;
++ (instancetype)sharedInstance;
+- (instancetype)initWithTag:(NSString *)tag;
+- (void)logErrorWithFormatArgs:(NSArray *)args andStackTrace:(NSArray<NSString*>*)stackTrace;
+- (void)logWithType:(MOLoggerType)type andFormatArgs:(NSArray *)args;
+- (BOOL)clearAllLogs;
+- (void)saveLogsToFile:(BOOL)save;
+- (void)enableLogs:(BOOL)enable;
+- (void)setLogFilesDuration:(NSInteger) duration;
 
 @end
+
+NS_ASSUME_NONNULL_END