+++ Frameworks/MOInApp.framework/CHANGELOG.md
@@ -2,6 +2,11 @@
 =====================
 ## Change Log:
 
+### Version 2.0.0  *(18th January, 2021)*
+-------------------------------------------
+* Changes to support  MoEngage-iOS-SDK 7.0.0
+* InApp exit animation fixes.
+
 ### Version 1.1.0  *(17th December, 2020)*
 -------------------------------------------
 * Changes to support  `enableSDK()` and `disableSDK()` methods for blocking/unblocking SDK features